Some missing cleanups
